Work in progress for Interoperability module
The Interoperability module is a reworked version of the future deprecated BindingsManager with the objective of offering better, cached and reliable interoperability functions and models to create a better bridge between the JavaScript interface and the .NET CLR. It was made taking all the current features that can be used in the current CLR in consideration.

Future support to manipulate and fetch generic types will also be added to Interoperability method.